refactor: Pull out test fixture to default location
To be able to share between tests.
